Refactor entrypoint into modular state, render, and controls
