+++ index.html
@@ -91,7 +91,7 @@
   <link rel="stylesheet" href="style.css" />
   <link rel="stylesheet" href="style-print.css" media="print" />
   <link rel="stylesheet" href="print.css" />
-  <script type="module" src="js/main.js"></script>
+  <script type="module" defer src="js/controls.js"></script>
 </head>
 <body>
   <main class="container py-4">
